ZSH QoL: Add make -j12 alias, DEBUGFS_PAGER=cat and redstderr
aliasing make to 'make -j12'

exporing DEBUGFS_PAGER=cat, since debugfs uses less for the output of
every. damn. command.

redstderr is a fun function that can be prefixed before a command. It
causes the stderr output of that command to be marked red, thus making
it easier to distinguish between stdout and stderr output.